added protection for prod in custom commands

+++ server/goalserver/planapp/management/commands/delete.py
@@ -1,5 +1,6 @@
 from django.core.management.base import BaseCommand, CommandError
 from django.contrib.auth.models import AnonymousUser, User
+from django.conf import settings
 from planapp.models import Goal, Plan, Task
 import datetime
 
@@ -11,8 +12,10 @@ class Command(BaseCommand):
         pass
 
     def handle(self, *args, **options):
-        for u in User.objects.all():
-            u.delete()
-
+        if not settings.PROD:
+            for u in User.objects.all():
+                u.delete()
         self.stdout.write(self.style.SUCCESS(str(datetime.datetime.now())))
         self.stdout.write(self.style.SUCCESS("deleted all data"))
+        else:
+            self.stdout.write(self.style.FAILURE("cannot run delete on prod"))
